Revert "Hr"
