+++ .eslintrc.cjs
@@ -16,6 +16,5 @@ module.exports = {
       'warn',
       { allowConstantExport: true },
     ],
-    'no-unused-vars': 'warn',
   },
-};
+}
