+++ src/MainPage.jsx
@@ -1,43 +1,68 @@
 import { useState, useEffect } from 'react';
-import { ReactDOM } from 'react';
 import { fabric } from 'fabric';
 
 import { 
-  // Header,
-  // Body,
+  Header,
+  Body,
   Container, 
-  // LeftContainer, 
+  LeftContainer, 
   Footer, 
-  // HeaderForeground,
-  // NavigationBar, 
-  // Nav1,
-  // Nav2,
-  // CanvasSpace,
-} from './components/MainPageStyles';
-import { NavigatorBar } from './components/NavigatorBar';
+  HeaderForeground,
+  NavigationBar, 
+  Nav1,
+  Nav2,
+  CanvasSpace,} from './components/MainPageStyles';
+import { ButtonGroupContainer } from './components/ButtonGroupContainer';
 
 function App() {
+  const [canvas, setCanvas] = useState(null);
+
+  useEffect(() => {
+    const initCanvas = () =>
+      new fabric.Canvas('canvas', {
+        height: 500,
+        width: 415,
+        backgroundColor: 'white',
+      });
+
+    setCanvas(initCanvas());
+  }, []);
+
+  useEffect(() => {
+    if (canvas) {
+      // 생성한 Canvas 인스턴스를 DOM에 추가
+      canvas.initialize('canvas');
+    }
+  }, [canvas]);
+
+  
 
   return (
     <>
     <Container>
-      <NavigatorBar/>
-      {/* <Header>
-        <NavigatorBar/>
-      {/*         
-        <NavigationBar className='nav-menu'>
+      <Header>
+        
+        <NavigationBar>
           <Nav1>
-            Make Own Polaroid!
+            Make your Polaroid!
           </Nav1>
           <Nav2>
             What is Y.O.P?
           </Nav2>
         </NavigationBar>
-           */}
+          
         
-        {/* <HeaderForeground>
-        </HeaderForeground> */}
-      {/* </Header> */}
+        <HeaderForeground>
+        </HeaderForeground>
+      </Header>
+      <Body>
+        <LeftContainer>
+          <ButtonGroupContainer />
+          <CanvasSpace>
+            <canvas id="canvas" />
+          </CanvasSpace>
+        </LeftContainer>
+      </Body>
       <Footer>
         
       </Footer>
