--- src/components/NavigatorBar/NavigatorBar.css
@@ -1,64 +1,0 @@
-.nav-tabs:hover {
-  background-color: #7198FF;
-}
-
-.active-nav-tabs:hover {
-  background-color: #FFFFFF
-}
-
-.nav-tabs {
-  cursor: pointer;
-  width: 12em;
-  height: 50%;
-  border-color: #FFFFFF;
-  border-top-left-radius: 15px;
-  border-top-right-radius: 15px;
-  background-color: #E0E0E0;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  /* align-content: top; */
-  text-align: center;
-  box-shadow: 5px 3px 2px 0.1px rgb(0, 0, 0, 0.1);
-  flex-flow: column;
-  overflow: hidden;
-  /* flex: 0; */
-}
-
-.active-nav-tabs {
-  height: 100%;
-  background-color: #FFFFFF;
-  display: flex;
-  text-align: center;
-  box-shadow: 5px 3px 2px 0.1px rgb(0, 0, 0, 0.1);
-  z-index: 1;
-}
-
-/* .active-nav-tabs::before {
-  width: auto;
-  height: auto;
-  border-color: #FFFFFF;
-  border-top-left-radius: 15px;
-  border-top-right-radius: 15px;
-  background-color: #FFFFFF;
-  display: flex;
-  align-items: start;
-  text-align: center;
-  box-shadow: 5px 2px 2px 1px rgb(0, 0, 0, 0.2);
-} */
-
-.content {
-
-  display: none;
-  background-color: #FFFFFF;
-  padding-top: 2%;
-  padding-bottom: 2%;
-
-}
-
-.active-content {
-  display: block;
-  background-color: #FFFFFF;
-  padding-top: 2%;
-  padding-bottom: 2%;
-}
